disparity_refinement/pyproject.toml
[project]
name = "disparity-refinement"
version = "0.1.0"
description = ""
authors = [
    {name = "Rohan",email = "rohankhaire.work@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch (>=2.7.0,<3.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
